# public/swagger.json
{
	"components": {
...
		"securitySchemes": {}
	},
	"info": {
		"title": "course_mern",
		"version": "1.0.0",
		"description": "MERN  Mongodb + Express + React + Node  -  Course in OpenBootcamp",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Leonardo Torrealba"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/katas/{{id}}": {
			"get": {
				"operationId": "GetKataById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to retrieve one kata that your id have been equal to id in param",
				"tags": [
					"KatasController"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}